+++ api/schedules/status_plots.py
@@ -14,20 +14,23 @@ from api import app
 from api.commands import chia_cli
 from api import utils
 
-first_run = True
+# Due to load, only send full plots list every X minutes
+FULL_SEND_INTERVAL_MINS = 15
+
+last_full_send_time = None
 
 def update():
-    global first_run
+    global last_full_send_time
     if not globals.farming_enabled() and not globals.harvesting_enabled():
         #app.logger.info("Skipping plotting status collection on farming/harvesting instance.")
         return
     with app.app_context():
         try:
-            if first_run:  # On first launch, send all plots
-                first_run = False
-                since = None
-            else: # On subsequent schedules, load only plots new in last 15 minutes.
-                since = (datetime.datetime.now() - datetime.timedelta(minutes=15)).strftime("%Y-%m-%d %H:%M:%S.000")
+            since = (datetime.datetime.now() - datetime.timedelta(minutes=FULL_SEND_INTERVAL_MINS)).strftime("%Y-%m-%d %H:%M:%S.000")
+            if not last_full_send_time or last_full_send_time <= \
+                (datetime.datetime.now() - datetime.timedelta(minutes=FULL_SEND_INTERVAL_MINS)):
+                since = None  # No since filter sends all plots, not just recent
+                last_full_send_time = datetime.datetime.now()
             hostname = utils.get_hostname()
             plots_farming = chia_cli.load_plots_farming()
             plots_by_id = {}