+++ scripts/dev/start-web.sh
@@ -23,4 +23,5 @@ fi
     --log-level=$LOG_LEVEL \
     --workers=2 \
     --access-logfile - \
+    --error-logfile - \
     web:app