+++ entrypoint.sh
@@ -10,6 +10,9 @@ if [[ ${blockchains} =~ flax ]]; then
   /usr/bin/bash /machinaris/scripts/flax_launch.sh
 fi
 
+# Build bladebit plotter on first run of container
+/usr/bin/bash /machinaris/scripts/bladebit_make.sh
+
 # Launch Machinaris web server and other services
 /machinaris/scripts/start_machinaris.sh